requirements.txt
Flask
Flask-WTF
Flask-Uploads
Pillow
WTForms

